+++ packages/core/test/files/FilesLoaders.spec.js
@@ -143,6 +143,14 @@ describe("FilesLoaders", () => {
     });
   });
 
+  describe("loaders getter", () => {
+    it("should return loaders", async () => {
+      await filesLoader.init();
+      expect(filesLoader.loaders.collections).toBeDefined();
+      expect(filesLoader.loaders.routes).toBeDefined();
+    });
+  });
+
   describe("when initialized", () => {
     it("should not require files from mocks folders if it is disabled", async () => {
       enabledOption.value = false;
@@ -162,7 +170,7 @@ describe("FilesLoaders", () => {
     it("should not throw and add an alert if there is an error loading route files", async () => {
       libsMocks.stubs.globule.find.returns(["foo"]);
       await filesLoader.init();
-      expect(alerts.flat.length).toEqual(2);
+      expect(alerts.flat.length).toEqual(1);
     });
 
     it("should add an alert when a routes file content does not pass validation", async () => {
@@ -176,10 +184,28 @@ describe("FilesLoaders", () => {
       filesLoader._babelRegisterOption = babelRegisterOption;
       filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
       await filesLoader.init();
-      expect(alerts.flat[1].collection).toEqual("files:routes:file");
-      expect(alerts.flat[1].id).toEqual("foo");
-      expect(alerts.flat[1].value.message).toEqual(
+      expect(alerts.flat[0].collection).toEqual("files:loader:routes:file");
+      expect(alerts.flat[0].id).toEqual("foo");
+      expect(alerts.flat[0].value.message).toEqual(
         expect.stringContaining("Error loading routes from file foo")
+      );
+    });
+
+    it("should add an alert when collections file content does not pass validation", async () => {
+      libsMocks.stubs.globule.find.returns(["collections.json"]);
+      filesLoader = new FilesLoaders(pluginMethods, {
+        requireCache,
+        require: () => ({}),
+      });
+      filesLoader._pathOption = pathOption;
+      filesLoader._watchOption = watchOption;
+      filesLoader._babelRegisterOption = babelRegisterOption;
+      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await filesLoader.init();
+      expect(alerts.flat[0].collection).toEqual("files:loader:collections");
+      expect(alerts.flat[0].id).toEqual("error");
+      expect(alerts.flat[0].value.message).toEqual(
+        expect.stringContaining("Error loading collections from file")
       );
     });
 
@@ -202,7 +228,8 @@ describe("FilesLoaders", () => {
       expect(alerts.flat.length).toEqual(1);
     });
 
-    it("should remove alerts when mocks file loads successfully", async () => {
+    it("should remove alerts when collections file loads successfully", async () => {
+      libsMocks.stubs.globule.find.returns(["foo"]);
       filesLoader = new FilesLoaders(pluginMethods, {
         requireCache,
         require: () => [],
@@ -213,6 +240,23 @@ describe("FilesLoaders", () => {
       filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
       await filesLoader.init();
       expect(alerts.flat.length).toEqual(0);
+    });
+
+    it("should add an alert when no collections file is found", async () => {
+      libsMocks.stubs.globule.find.returns([]);
+      filesLoader = new FilesLoaders(pluginMethods, {
+        requireCache,
+        require: () => [],
+      });
+      filesLoader._pathOption = pathOption;
+      filesLoader._watchOption = watchOption;
+      filesLoader._babelRegisterOption = babelRegisterOption;
+      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await filesLoader.init();
+      expect(alerts.flat.length).toEqual(1);
+      expect(alerts.flat[0].value.message).toEqual(
+        expect.stringContaining("No collections file was found")
+      );
     });
 
     it("should call to loadCollections method when mocks file is loaded", async () => {
@@ -228,39 +272,39 @@ describe("FilesLoaders", () => {
       expect(pluginMethods.loadCollections.callCount).toEqual(1);
     });
 
-    it("should try to load collections.json when collections.js file does not exists", async () => {
-      filesLoader = new FilesLoaders(pluginMethods, {
-        requireCache,
-        require: sandbox.spy,
-      });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      libsMocks.stubs.fsExtra.existsSync.onCall(0).returns(true);
-      libsMocks.stubs.fsExtra.existsSync.onCall(1).returns(false);
-      libsMocks.stubs.fsExtra.existsSync.onCall(2).returns(true);
-      await filesLoader.init();
-      expect(libsMocks.stubs.fsExtra.existsSync.callCount).toEqual(3);
-    });
-
     it("should add an alert when file name is mocks", async () => {
+      libsMocks.stubs.globule.find.returns(["mocks.json"]);
       filesLoader = new FilesLoaders(pluginMethods, {
         requireCache,
-        require: sandbox.spy,
+        require: () => [],
       });
       filesLoader._pathOption = pathOption;
       filesLoader._watchOption = watchOption;
       filesLoader._babelRegisterOption = babelRegisterOption;
       filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      libsMocks.stubs.fsExtra.existsSync.onCall(0).returns(true);
-      libsMocks.stubs.fsExtra.existsSync.onCall(1).returns(false);
-      libsMocks.stubs.fsExtra.existsSync.onCall(2).returns(false);
-      libsMocks.stubs.fsExtra.existsSync.onCall(3).returns(true);
       await filesLoader.init();
       expect(alerts.flat[0].value.message).toEqual(
         "Defining collections in 'mocks.json' file is deprecated. Please rename it to 'collections.json'"
       );
+    });
+
+    it("should remove alert when file name is collections", async () => {
+      libsMocks.stubs.globule.find.returns(["mocks.json"]);
+      filesLoader = new FilesLoaders(pluginMethods, {
+        requireCache,
+        require: () => [],
+      });
+      filesLoader._pathOption = pathOption;
+      filesLoader._watchOption = watchOption;
+      filesLoader._babelRegisterOption = babelRegisterOption;
+      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await filesLoader.init();
+      expect(alerts.flat[0].value.message).toEqual(
+        "Defining collections in 'mocks.json' file is deprecated. Please rename it to 'collections.json'"
+      );
+      libsMocks.stubs.globule.find.returns(["collections.json"]);
+      filesLoader._loadFiles();
+      expect(alerts.flat.length).toEqual(0);
     });
 
     it("should not throw and add an alert if there is an error in loadRoutesfiles method", async () => {