+++ packages/main/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Breaking change
 
+## [unreleased]
+
+- feat: Trace start error
+
 ## [3.8.0] - 2022-08-04
 
 ### Changed